Auto-generate SocketListener::accept_async() and related functions
The annotations were fixed long ago, as was the one for
ThreadedSocketService::run().
