+++ src/prelude.rs
@@ -28,7 +28,6 @@ pub use pollable_input_stream::PollableInputStreamExtManual;
 pub use pollable_output_stream::PollableOutputStreamExtManual;
 pub use settings::SettingsExtManual;
 pub use socket::*;
-pub use socket_listener::SocketListenerExtManual;
 #[cfg(any(unix, feature = "dox"))]
 pub use unix_input_stream::UnixInputStreamExtManual;
 #[cfg(any(unix, feature = "dox"))]
